Add some material on replacing scanners and vars
Add a recipe on replacing a scanner (as long as it has a directory
as the function) - mostly as a place to store information discovered
about switching to the C Conditional Scanner.

Add a recipe on replacing construction variable contents when you
want to do a partial update, not just a full replacement.

Two recipes have small updates.

Signed-off-by: Mats Wichmann <mats@linux.com>